refactor(core): share JSX message assembly
